WIP: Initial pass at adding hikey960_gki fragment
To try to make it easier to stay in sync w/ gki_defconfig switch
to using a hikey960_gki.fragment file which we use merge_config
to merge on top of gki_defconfig.

For sanity's sake, we'll keep the resulting generated config as
hikey960_gki_defconfig, so that its easier to notice changes
to gki_defconfig that break things.

Change-Id: Ie88805656e5e982aee752be76d0f2e83f0f69fdd
Signed-off-by: John Stultz <john.stultz@linaro.org>