drm: kirin: Tweak the LDI handling
With the change ff57c6513820e ("drm: kirin: Fix for hikey620
display offset problem") we added support for handling LDI
overflows by reseting the hardware.

However, its been observed that when we do hit the LDI overflow
condition, the irq seems to be screaming, and we do nothing but
stream:
[drm:ade_irq_handler [kirin_drm]] *ERROR* LDI underflow!
over and over to the screen.

This patch tries to tweak the LDI handling so we don't ack the
irq until after the reset has occured. It also removes the error
message printing to avoid slowing down the handling of the
issue.

Change-Id: Iae43e7f4bebad06d9ef2bdcfa70b43e38e7819e3
Fixes: ff57c6513820e ("drm: kirin: Fix for hikey620 display offset problem")
Signed-off-by: John Stultz <john.stultz@linaro.org>