block: blk-crypto-fallback for Inline Encryption
Blk-crypto delegates crypto operations to inline encryption hardware when
available. The separately configurable blk-crypto-fallback contains a
software fallback to the kernel crypto API - when enabled, blk-crypto
will use this fallback for en/decryption when inline encryption hardware is
not available. This lets upper layers not have to worry about whether or
not the underlying device has support for inline encryption before
deciding to specify an encryption context for a bio. It also allows for
testing without actual inline encryption hardware - in particular, it
makes it possible to test the inline encryption code in ext4 and f2fs
simply by running xfstests with the inlinecrypt mount option, which in
turn allows for things like the regular upstream regression testing of
ext4 to cover the inline encryption code paths. For more details, refer
to Documentation/block/inline-encryption.rst.

Signed-off-by: Satya Tangirala <satyat@google.com>
Signed-off-by: John Stultz <john.stultz@linaro.org>